+++ sdk/cosmos/azure-cosmos-test/src/main/java/com/azure/cosmos/test/implementation/faultinjection/FaultInjectionRuleProcessor.java
@@ -326,7 +326,6 @@ public class FaultInjectionRuleProcessor {
             case READ_ITEM:
             case METADATA_REQUEST_CONTAINER:
             case METADATA_REQUEST_DATABASE_ACCOUNT:
-            case METADATA_REQUEST_ADDRESS_REFRESH:
                 return OperationType.Read;
             case CREATE_ITEM:
                 return OperationType.Create;
@@ -348,6 +347,8 @@ public class FaultInjectionRuleProcessor {
                 return OperationType.QueryPlan;
             case METADATA_REQUEST_PARTITION_KEY_RANGES:
                 return OperationType.ReadFeed;
+            case METADATA_REQUEST_ADDRESS_REFRESH: // address refresh can happen for any operations: read, write, query etc
+                return null;
             default:
                 throw new IllegalStateException("FaultInjectionOperationType " + faultInjectionOperationType + " is not supported");
         }
@@ -487,8 +488,12 @@ public class FaultInjectionRuleProcessor {
     }
 
     private boolean isWriteOnly(FaultInjectionCondition condition) {
-        return condition.getOperationType() != null
-            && this.getEffectiveOperationType(condition.getOperationType()).isWriteOperation();
+        if (condition.getOperationType() != null) {
+            OperationType effectiveOperationType = this.getEffectiveOperationType(condition.getOperationType());
+            return effectiveOperationType != null && effectiveOperationType.isWriteOperation();
+        }
+
+        return false;
     }
 
     static class FaultInjectionRuleProcessorRetryPolicy implements IRetryPolicy {